+++ Helper/Markdown.php
@@ -16,7 +16,10 @@ class Markdown extends \Lime\Helper {
             'cached_toc_format' => 'flat',
         ];
 
-        $config = \array_replace_recursive($default, $this->app->retrieve('config/bettermarkdown', []));
+        $config = \array_replace_recursive(
+            $default,
+            $this->app->retrieve('config/bettermarkdown', [])
+        );
 
         $this->config = $config;
 
@@ -32,8 +35,12 @@ class Markdown extends \Lime\Helper {
 
             case 'extended':
             default:
-                $this->parser = new ParsedownCheckbox();
-                $this->parser->options = \array_merge($this->parser->options, $config['toc'] ?? []);
+                $this->parser = new \ParsedownTasks();
+                $this->parser->options = \array_merge(
+                    $this->parser->options ?? [],
+                    $config['toc'] ?? [],
+                    $config['tasks'] ?? []
+                );
                 break;
 
         }
@@ -157,124 +164,4 @@ class Markdown extends \Lime\Helper {
         return $tmp['children'];
     }
 
-}
-
-// Because of the weird extension behaviour of Parsedown, I had to copy/paste the original ParsedownCheckbox
-// class with a namespace modification. So running `composer update` will not effect that class without
-// manual adjustments.
-// source: https://github.com/leblanc-simon/parsedown-checkbox
-// author: Simon Leblanc, MIT Licensed
-
-use \ParsedownToC as ParsedownExtra;
-
-/**
- * This file is part of the ParsedownCheckbox package.
- *
- * (c) Simon Leblanc <contact@leblanc-simon.eu>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-class ParsedownCheckbox extends ParsedownExtra
-{
-    const VERSION = '0.2.0';
-
-    public function __construct()
-    {
-        parent::__construct();
-
-        array_unshift($this->BlockTypes['['], 'Checkbox');
-    }
-
-    protected function blockCheckbox($line)
-    {
-        $text = trim($line['text']);
-        $begin_line = substr($text, 0, 4);
-        if ('[ ] ' === $begin_line) {
-            return [
-                'handler' => 'checkboxUnchecked',
-                'text' => substr(trim($text), 4),
-            ];
-        }
-
-        if ('[x] ' === $begin_line) {
-            return [
-                'handler' => 'checkboxChecked',
-                'text' => substr(trim($text), 4),
-            ];
-        }
-    }
-
-    protected function blockListComplete(array $block)
-    {
-        foreach ($block['element']['elements'] as &$li_element) {
-            foreach ($li_element['handler']['argument'] as $text) {
-                $begin_line = substr(trim($text), 0, 4);
-                if ('[ ] ' === $begin_line) {
-                    $li_element['attributes'] = ['class' => 'parsedown-task-list parsedown-task-list-open'];
-                } elseif ('[x] ' === $begin_line) {
-                    $li_element['attributes'] = ['class' => 'parsedown-task-list parsedown-task-list-close'];
-                }
-            }
-        }
-
-        return $block;
-    }
-
-    protected function blockCheckboxContinue(array $block)
-    {
-    }
-
-    protected function blockCheckboxComplete(array $block)
-    {
-        $block['element'] = [
-            'rawHtml' => $this->{$block['handler']}($block['text']),
-            'allowRawHtmlInSafeMode' => true,
-        ];
-
-        return $block;
-    }
-
-    protected function checkboxUnchecked($text)
-    {
-        if ($this->markupEscaped || $this->safeMode) {
-            $text = self::escape($text);
-        }
-
-        return '<input type="checkbox" disabled /> ' . $this->format($text);
-    }
-
-    protected function checkboxChecked($text)
-    {
-        if ($this->markupEscaped || $this->safeMode) {
-            $text = self::escape($text);
-        }
-
-        return '<input type="checkbox" checked disabled /> ' . $this->format($text);
-    }
-
-    /**
-     * Formats the checkbox label without double escaping.
-     * @param string $text the string to format
-     * @return string the formatted text
-     */
-    protected function format($text)
-    {
-        // backup settings
-        $markup_escaped = $this->markupEscaped;
-        $safe_mode = $this->safeMode;
-
-        // disable rules to prevent double escaping.
-        $this->setMarkupEscaped(false);
-        $this->setSafeMode(false);
-
-        // format line
-        $text = $this->line($text);
-
-        // reset old values
-        $this->setMarkupEscaped($markup_escaped);
-        $this->setSafeMode($safe_mode);
-
-        return $text;
-    }
 }