+++ bootstrap.php
@@ -2,12 +2,12 @@
 /**
  * Cache, task lists and ToC support for markdown conversion with Cockpit CMS v1
  *
- * @see       https://github.com/raffaelj/cockpit_BetterMarkdown
- * @see       https://github.com/agentejo/cockpit/
- *
- * @version   0.1.1
+ * @version   0.2.0
  * @author    Raffael Jesche
  * @license   MIT
+ *
+ * @see       https://github.com/raffaelj/cockpit_BetterMarkdown
+ * @see       https://github.com/agentejo/cockpit/
  */
 
 require_once(__DIR__.'/lib/vendor/autoload.php');